app/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.82.0
python-dotenv>=1.1.0
pydantic>=2.11.0
sqlalchemy>=2.0.0
alembic>=1.15.0
aiosqlite>=0.21.0
pyyaml>=6.0
cryptography>=45.0.0
